% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\docType{data}
\name{coins}
\alias{coins}
\title{Coin Tossing}
\format{
A data frame with 13 observations of 3 variables
\describe{
  \item{id}{participant number}
  \item{heads}{number of heads tossed}
  \item{tails}{number of tails tossed}
}

From the authors: "The toss of a coin has been a method used to determine random outcomes for centuries. It is still used in some research studies as a method of randomization, although it has largely been discredited as a valid randomization method. We sought to provide evidence that the toss of a coin can be manipulated."
}
\source{
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2789164/
}
\usage{
coins
}
\description{
Med school students trying to toss heads.
}
\details{
Thirteen participants each tossed a coin 300 times.
}
\keyword{datasets}
