% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_fatalities.R
\docType{data}
\name{driver_fatalities}
\alias{driver_fatalities}
\title{Traffic Fatalities}
\format{
A data frame with 103 observations of 54 variables
\describe{
  \item{cfips}{State County FIPS code}
  \item{statenum}{State FIPS}
  \item{state}{State name}
  \item{stateabbrev}{State abbreviation}
  \item{county}{County FIPS}
  \item{countyname}{County name}
  \item{cbsaname}{New MSA Name}
  \item{year}{Year}
  \item{month}{Month}
  \item{uber_active}{Uber active in county}
  \item{uber_or_lyft}{Number of rideshare active}
  \item{uberx_active}{UberX active in county}
  \item{fatal}{Traffic fatalities (M)}
  \item{drunkdrivera}{Fatal accidents in which there was a reported drunk driver}
  \item{drunkdriverb}{Fatalalties in which there was a reported drunk driver}
  \item{crashes}{Crashes with at least one fatality (M)}
  \item{weekend_holidaya}{Fatal accidents on weekend/holidays (M)}
  \item{weekend_holidayb}{Fatalalities on weekend/holidays (M)}
  \item{fatal2}{Log fatalities + 1}
  \item{fatal3}{Fatalities per population}
  \item{weekend_holidayb2}{Log Fatalalities on weekend/holidays + 1}
  \item{drunkdriverb2}{Log Fatalalties in which there was a reported drunk driver + 1}
  \item{weekend_holidayb3}{Fatalalities on weekend/holidays per population}
  \item{drunkdriverb3}{Fatalalties in which there was a reported drunk driver per population}
  \item{pop}{Total county population (Yearly)}
  \item{vmt_state}{State Vehicle Miles Travelled (Monthly)}
  \item{lengthp2008}{County Percent of State Road Length in 2008}
  \item{vmt}{County Vehicle Miles Travelled (Monthly)}
  \item{unemp}{County Unemployment rate (M)}
  \item{bt}{State Beer tax rate (Y)}
  \item{taxi_county_rate}{taxi_msa_rate weighted by countymsa_pop (Y)}
  \item{decrim}{State Has marijuana decriminalization law (M)}
  \item{gdl}{State Has graduated driver-licensing law (M)}
  \item{hh}{State Has hands-free cell phone law (M)}
  \item{med_mar}{State Has medical marijuana legalization law (M)}
  \item{per_se}{State Has drug per se law (M)}
  \item{primsb}{State Has primary seat belt law (M)}
  \item{secsb}{State Has secondary seat belt law (M)}
  \item{txt_ban}{State Has cell phone texting ban law (M)}
  \item{tpop}{Total population year before Uber}
  \item{lmhh}{Log median household income year before Uber}
  \item{p2039}{Percent 20-30 yr old year before Uber}
  \item{col}{Percent college degree year before Uber}
  \item{vehhh}{Percent vehicle ownership year before Uber}
  \item{alc}{Percent excessive drinking year before Uber}
  \item{vmtt_rate}{Public Transportation VMT per Total VMT (Y)}
  \item{vmtr}{Vehicle rev miles travelled - Rail (Y)}
  \item{vmtr_rate}{Rail VMT per Total VMT (Y)}
  \item{cov_all}{Public transit coverage}
  \item{ser_all}{Public transit service frequency}
  \item{puapop}{Percent in census urban area}
  \item{popdensity}{Population density}
  \item{drinkpld}{Number of drinking establishments per area}
  \item{anyall}{Percent any drinking year before Uber}
}

This data is used to investigate the impact of Uber on trafic fatalities. From the authors: "Ridehailing services such as Uber have been promoted as viable interventions for curbing alcohol-involved driving fatalities. However, evidence of ridehailing’s impact has been mixed, with some studies finding no association but others finding either an increase or a decrease in fatalities."
}
\source{
https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0238744
}
\usage{
driver_fatalities
}
\description{
Number of traffic fatalities in urban areas in January, 2017.
}
\keyword{datasets}
