% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_temp.R
\docType{data}
\name{sex_temp}
\alias{sex_temp}
\title{Effect of temperature on performance by sex}
\format{
A data frame with 543 observations of 30 variables.
\describe{
  \item{session}{There were 24 sessions consisting of 14-25 tests.}
  \item{temp}{Temperature of room in degrees Celsius.}
  \item{crt}{Participants had 5 minutes to answer three original CRT questions introduced by S. Frederick. Number of correct answers reported.}
  \item{mathreported}{Number of math questions answered.}
  \item{mathcorrect}{Number of math questions correctly answered.}
  \item{mathmistakes}{Number of math questions incorrectly answered.}
  \item{wordstotalreported}{Number of words (in German) built from letters ADEHINRSTU in five mminutes. Rewards for longer words.}
  \item{wordstotalcorrect}{Number of words that were correctly built.}
  \item{wordsmistakes}{Number of words that were incorrect.}
  \item{repeats}{Not described in the paper. Possibly number of words that were repeated.}
  \item{word[2-9]}{Not described. Probably the lengths of the correct words, where wordx contains x letters.}
  \item{male}{1 if male, 0 if female. Experiment only allowed for those choices.}
  \item{age}{Self-reported age of participant.}
  \item{majorecon}{1 if econ major, 0 if not an econ major.}
  \item{nativegerman}{1 if native German speaker}
  \item{enjoymath}{On scale from 0 to 10, how much did you enjoy math task?}
  \item{enjoywords}{On scale from 0 to 10, how much did you enjoy the words task?}
  \item{strongme-month}{None of these columns seem to be described in paper.}
}
}
\source{
Chang TY, Kajackaite A (2019) Battle for the thermostat: Gender and the effect of temperature on cognitive performance. PLoS ONE 14(5): e0216362. https://doi.org/10.1371/journal.pone.0216362
}
\usage{
sex_temp
}
\description{
Results from test given at various temperature to males and females.
}
\details{
From the authors: "The fact that women generally prefer higher indoor temperatures than men is well supported by survey evidence. This difference in preferences, sometimes referred to as the “battle of the thermostat,” is passionately discussed in popular culture and has received considerable media attention. Surprisingly then, the research examining the impact of temperature on cognitive performance has not explored a link between gender and temperature response."
}
\keyword{datasets}
